Make TT only store 16-bit keys
Bench: 9260314

+++ engine/ttable.hpp
@@ -15,7 +15,7 @@ enum TTFlag {
 
 struct TTable {
 	struct TTEntry {
-		uint32_t key;
+		uint16_t key;
 		Move best_move;
 		Value eval, s_eval;
 		uint8_t depth;
